feat: beta icon export
